// src/types/ThermostatSchema.ts
/* tslint:disable */
/* eslint-disable */
/**
 * This file is generated by core-types-ts on behalf of typeconv, DO NOT EDIT.
 * For more information, see:
 *  - {@link https://github.com/grantila/core-types-ts}
 *  - {@link https://github.com/grantila/typeconv}
 */

/** Any smart thermostat */
export interface ThermostatSchema {
    id: string;
    name?: string;
    /**
     * @default
     *     thermostat
     */
    type: "alarm" | "dimmer" | "switch" | "motionSensor" | "windowCovering" | "camera" | "mediaSource" | "thermostat" | "lock" | "courtesy" | "gateway" | "tv" | "dvr" | "appleTv" | "discPlayer" | "mediaPlayer" | "uncontrolledDevice";
    driver: "aws-kinesis" | "butler" | "crestron" | "dmp" | "dormakaba" | "dsc" | "ecobee" | "igor" | "inncom" | "kohost-k7" | "kohost-pms" | "lg" | "lirc" | "mews" | "mht" | "paxton" | "pelican-wireless" | "power-shades" | "rebrandly" | "salto" | "salto-irn" | "se" | "sendgrid" | "sonifi" | "stay-n-touch" | "twilio" | "cloudflare-images" | "cloudflare-stream" | "insperia-privacy";
    offline?: boolean;
    supportedNotifications?: ("button 1" | "button 2" | "button 3" | "button 4" | "button 5" | "idle" | "powerHasBeedApplied" | "acMainsDisconnected" | "acMainsReconnected" | "replaceBatterySoon" | "replaceBatteryNow" | "hardwareFailure" | "softwareFailure" | "hardwareFailureWithCode" | "softwareFailureWithCode" | "motionDetection" | "airFilterNeedsCleaned" | "airFilterNeedsReplaced" | "smokeDetected" | "outsideSafeTemperatureRange" | "outsideSafeHumidityRange" | "scheduleMaintenance")[];
    notification?: {
        name?: "button 1" | "button 2" | "button 3" | "button 4" | "button 5" | "idle" | "powerHasBeedApplied" | "acMainsDisconnected" | "acMainsReconnected" | "replaceBatterySoon" | "replaceBatteryNow" | "hardwareFailure" | "softwareFailure" | "hardwareFailureWithCode" | "softwareFailureWithCode" | "motionDetection" | "airFilterNeedsCleaned" | "airFilterNeedsReplaced" | "smokeDetected" | "outsideSafeTemperatureRange" | "outsideSafeHumidityRange" | "scheduleMaintenance";
        timestamp?: number;
        description?: string;
        [key: string]: any;
    } | null;
    currentTemperature?: number;
    currentHumidity?: number;
    hvacMode: "cool" | "heat" | "auto" | "off";
    hvacState: ("cooling" | "heating" | "off") | null;
    fanMode: "auto" | "low" | "medium" | "high" | "off" | "on";
    fanState: ("off" | "low" | "medium" | "high" | "on") | null;
    /**
     * @default
     *     fahrenheit
     */
    temperatureScale: "celsius" | "fahrenheit";
    humidityScale?: ("absolute" | "relative") | null;
    supportedHvacModes: ("cool" | "heat" | "auto" | "off")[];
    supportedFanModes: ("auto" | "low" | "medium" | "high" | "off" | "on")[];
    setpoints: {
        cool?: {
            value?: number;
            min?: number | null;
            max?: number | null;
        };
        heat?: {
            value?: number;
            min?: number | null;
            max?: number | null;
        };
        auto?: {
            value?: number;
            min?: number | null;
            max?: number | null;
        };
    };
    /**
     * @default
     *     3
     */
    minAutoDelta?: number;
    cycleRate?: number;
    batteryLevel?: number;
    systemId?: string;
    watts?: number;
    [key: string]: any;
}